requirements.txt
mcp
mysql-connector-python
python-dotenv
openai